backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.1.2
beanie==1.25.0
beautifulsoup4==4.12.3
black==24.10.0
bs4==0.0.2
cattrs==23.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
ell-ai==0.0.14
email-validator==2.2.0
exceptiongroup==1.2.0
farama-notifications==0.0.4
fastapi==0.111.1
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
groq==0.11.0
gymnasium==1.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
iniconfig==2.0.0
jinja2==3.1.4
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain>=0.1.0
langchain-community>=0.0.38
langchain-core>=0.1.27
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
lazy-model==0.2.0
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mongomock==4.1.2
mongomock-motor==0.0.29
motor==3.4.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.54.4
opencv-python==4.10.0.84
orjson==3.10.11
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pluggy==1.4.0
propcache==0.2.0
psutil==5.9.8
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.0
pygame==2.6.1
pygments==2.18.0
pyjwt==2.8.0
pymongo==4.6.3
pyparsing==3.2.0
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sentinels==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.36
sqlmodel==0.0.21
starlette==0.37.2
swig==4.2.1.post0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tqdm==4.67.0
typer==0.13.0
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
yarl==1.17.2
